# docs/openapi.yaml
components:
    schemas:
...
        dto.CreateAIChatResponse:
            properties:
                created_at:
                    type: string
                deleted_at:
                    type: string
                id:
                    type: string
                title:
                    type: string
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
...
        dto.ErrorItem:
            properties:
                name:
                    type: string
                reason:
                    type: string
            type: object
        dto.ErrorResponse:
            properties:
                errors:
                    items:
                        $ref: '#/components/schemas/dto.ErrorItem'
                    type: array
                message:
                    type: string
            type: object
...
        dto.ListAIChatMessagesAndAnswersResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/entity.AIChatMessageAndAnswer'
                    type: array
                page:
                    type: integer
                page_size:
                    type: integer
                total_items:
                    type: integer
                total_pages:
                    type: integer
            type: object
        dto.ListAIChatsResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/entity.AIChat'
                    type: array
                page:
                    type: integer
                page_size:
                    type: integer
                total_items:
                    type: integer
                total_pages:
                    type: integer
            type: object
...
        dto.SignInRequest:
            properties:
                provider:
                    allOf:
                        - $ref: '#/components/schemas/entity.Provider'
                    enum:
                        - GOOGLE
                        - APPLE
                        - REFRESH
                        - MOCK
            required:
                - provider
            type: object
        dto.SignInResponse:
            properties:
                access_token:
                    type: string
                refresh_token:
                    type: string
                user:
                    $ref: '#/components/schemas/entity.User'
            type: object
...
        dto.UpdateAIChatRequest:
            properties:
                title:
                    type: string
            required:
                - title
            type: object
...
        entity.AIChat:
            properties:
                created_at:
                    type: string
                deleted_at:
                    type: string
                id:
                    type: string
                title:
                    type: string
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
        entity.AIChatMessageAndAnswer:
            properties:
                author:
                    $ref: '#/components/schemas/entity.AIChatMessageAuthor'
                created_at:
                    type: string
                id:
                    type: string
                message:
                    type: string
                rating:
                    type: string
            type: object
        entity.AIChatMessageAuthor:
            enum:
                - USER
                - ARTIFICIAL_INTELLIGENCE
            type: string
            x-enum-varnames:
                - AIChatMessageAuthorUser
                - AIChatMessageAuthorAI
...
        entity.Provider:
            enum:
                - GOOGLE
                - APPLE
                - REFRESH
                - MOCK
            type: string
            x-enum-varnames:
                - ProviderGoogle
                - ProviderApple
                - ProviderRefresh
                - ProviderMock
...
        entity.User:
            properties:
                avatar:
                    type: string
                created_at:
                    type: string
                deleted_at:
                    type: string
                email:
                    type: string
                id:
                    type: string
                name:
                    type: string
                subscription_expires_at:
                    type: string
                synchronized_at:
                    type: string
                tier:
                    type: string
                updated_at:
                    type: string
            type: object
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            description: Type "Bearer" followed by a space and JWT token.
            in: header
            name: Authorization
            type: apiKey
info:
    contact:
        email: danielmesquitta123@gmail.com
        name: Daniel Mesquita
    description: API Finance Manager
    title: API Finance Manager
    version: "1.0"
openapi: 3.0.3
paths:
...
    /v1/ai-chats:
        get:
            description: List ai chats
            parameters:
                - description: Search
                  in: query
                  name: search
                  schema:
                    type: string
                - description: Page
                  in: query
                  name: page
                  schema:
                    type: integer
                - description: Page size
                  in: query
                  name: page_size
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ListAIChatsResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List ai chats
            tags:
                - AI Chat
        post:
            description: Create ai chat
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.CreateAIChatResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Create ai chat
            tags:
                - AI Chat
    /v1/ai-chats/{ai_chat_id}:
        delete:
            description: Delete ai chat
            parameters:
                - description: AI Chat ID
                  in: path
                  name: ai_chat_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Delete ai chat
            tags:
                - AI Chat
        put:
            description: Update ai chat
            parameters:
                - description: AI Chat ID
                  in: path
                  name: ai_chat_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.UpdateAIChatRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Update ai chat
            tags:
                - AI Chat
    /v1/ai-chats/{ai_chat_id}/messages:
        get:
            description: List ai chats messages
            parameters:
                - description: AI Chat ID
                  in: path
                  name: ai_chat_id
                  required: true
                  schema:
                    format: uuid
                    type: string
                - description: Page
                  in: query
                  name: page
                  schema:
                    type: integer
                - description: Page size
                  in: query
                  name: page_size
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ListAIChatMessagesAndAnswersResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List ai chats messages
            tags:
                - AI Chat
        post:
            description: Generate ai chat message
            parameters:
                - description: AI Chat ID
                  in: path
                  name: ai_chat_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.UpdateAIChatRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.CreateAIChatResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Generate ai chat message
            tags:
                - AI Chat
    /v1/auth/refresh:
        post:
            description: Use refresh token to generate new access token
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.SignInResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Refresh token
            tags:
                - Auth
    /v1/auth/sign-in:
        post:
            description: Authenticate user through Google or Apple token
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.SignInRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.SignInResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Sign in
            tags:
                - Auth
...